api for login and registration

+++ src/main/java/com/example/Afternoon/Delights/entity/Member.java
@@ -1,10 +1,12 @@
 package com.example.Afternoon.Delights.entity;
 
+
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import lombok.Data;
+
 
 import java.time.LocalTime;
 
@@ -15,7 +17,7 @@ public class Member {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private Integer pin;
+    private Long pin;
 
     private String name;
